Slight restructure.
