+++ src/adapters/types/variants.ts
