+++ src/tooltips/structureTooltip.ts
